+++ tests/conftest.py
@@ -4,11 +4,12 @@ import numpy as np
 import pytest
 
 from qcio import (
-    CalcSpec,
-    CompositeCalcSpec,
+    DualProgramInput,
+    FileInput,
     Files,
-    FileSpec,
     OptimizationData,
+    ProgramArgs,
+    ProgramInput,
     Results,
     SinglePointData,
 )
@@ -29,7 +30,7 @@ def water():
 
 @pytest.fixture
 def file_input():
-    return FileSpec(
+    return FileInput(
         files={"binary": b"binary data", "text": "text data"},
         cmdline_args=["-i", "input.dat", "-o", "output.dat"],
     )
@@ -42,7 +43,7 @@ def input_data(request, file_input, calc_input, ccalc_input):
         return file_input
     elif request.param == "calc_input":
         return calc_input("energy")
-    elif request.param == "ccalc_input":  # CompositeCalcSpec
+    elif request.param == "ccalc_input":  # DualProgramInput
         return ccalc_input
     else:
         raise ValueError(f"Unknown input data type: {request.param}")
@@ -50,10 +51,10 @@ def input_data(request, file_input, calc_input, ccalc_input):
 
 @pytest.fixture
 def calc_input(water):
-    """Function that returns CalcSpec of calctype."""
+    """Function that returns ProgramInput of calctype."""
 
     def _create_calc_input(calctype):
-        return CalcSpec(
+        return ProgramInput(
             structure=water,
             calctype=calctype,
             model={"method": "hf", "basis": "sto-3g"},
@@ -71,10 +72,10 @@ def calc_input(water):
 
 @pytest.fixture
 def ccalc_input(water):
-    """Function that returns CompositeCalcSpec of calctype."""
+    """Function that returns DualProgramInput of calctype."""
 
     def _create_calc_input(calctype):
-        return CompositeCalcSpec(
+        return DualProgramInput(
             structure=water,
             calctype=calctype,
             keywords={
@@ -83,7 +84,7 @@ def ccalc_input(water):
                 "some-bool": False,
             },
             subprogram="fake subprogram",
-            subprogram_args={"model": {"method": "hf", "basis": "sto-3g"}},
+            subprogram_args=ProgramArgs(model={"method": "hf", "basis": "sto-3g"}),
         )
 
     return _create_calc_input
@@ -115,7 +116,7 @@ def results(calc_input, sp_data):
     pi_energy = calc_input("energy")
     sp_data = sp_data(pi_energy.structure)
 
-    return Results[CalcSpec, SinglePointData](
+    return Results[ProgramInput, SinglePointData](
         input_data=pi_energy,
         success=True,
         logs="program standard out...",
@@ -130,7 +131,7 @@ def results_failure(calc_input, sp_data):
     """Failed Results object"""
     ci_energy = calc_input("energy")
 
-    return Results[CalcSpec, Files](
+    return Results[ProgramInput, Files](
         input_data=ci_energy,
         success=False,
         traceback="Traceback...",